// src/services/pptx/types.ts
export interface HonorRollBackgrounds {
    bgCover: string
    bgQueens: string
    bgFirstPrincess: string
    bgSecondPrincess: string
}

export interface PointsClubBackgrounds {
    club300: string
    club600: string
    club900: string
    club1200: string
    club1500: string
    club2000: string
    club2500: string
    club3000: string
}

export interface RoadToSuccessBackgrounds {
    road_success_diq: string
    road_success_future_director: string
    road_success_target_a_diq: string
    road_success_target_a_fd: string
}

export interface StarsBackgrounds {
    diamond_star: string
    pearl_star: string
    emerald_star: string
    ruby_star: string
    sapphire_star: string
}

export interface GroupProductionBackgrounds {
    diamond_luminaire: string
    pearl_luminaire: string
    emerald_luminaire: string
    ruby_luminaire: string
    sapphire_luminaire: string
}

export interface PinkCircleBackgrounds {
    target_pink_circle: string
    pink_circle: string
    pink_circle_vip: string
    pink_circle_vip_gold: string
}